ml_requirements.txt
# ML Model Dependencies
# Install with: pip install -r ml_requirements.txt

# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0

# Image processing
Pillow>=10.0.0

# Optional: For training optimizations
# tensorboard>=2.13.0  # Training visualization
# albumentations>=1.3.0  # Advanced augmentations
# timm>=0.9.0  # Pre-trained models library
